refactor(contact): split contact model
